CheckAllowance intent now working

+++ setup.js
@@ -30,6 +30,7 @@ var accountId;
 program
     .version('1.0.0')
     .option('-t, --timetastic-token [value]')
+    .option('-i, --refresh-intents-only')
     .parse(process.argv);
 
 console.log("Holibot installer.");
@@ -63,76 +64,82 @@ function deploy(timeTasticToken) {
     fs.writeFileSync('token.yml', 'timetastic: ' + timeTasticToken);
 
     // execute each operation in sequence
-    async.waterfall([
-        // create the serverless project in AWS (i.e. the lambda function)
-        //deployServerlessProject,
+    async.waterfall(
+        []
 
-        // get the AWS ARN of the created lambda function
-        getLambdaFunctionArn]
+            // create the serverless project in AWS (i.e. the lambda function)
+            .concat([deployServerlessProject])
 
-        // delete any previously existing Holibot intents (we're going to recreate them)
-        .concat(
-        intents.map((intent) => (callback) => {
-            try {
-                Lex.deleteIntent({ name: intent.name }, (err, data) => callback(null));
-            }
-            catch (error) {
-                console.log('Could not delete ' + intent.name + ': ' + error);
-            }
-        })
-        )
+            // get the AWS ARN of the created lambda function
+            .concat([getLambdaFunctionArn])
 
-        // create each intent
-        .concat(
-        intents.map((intent) => (callback) => {
-            console.log("Creating intent: " + intent.name);
-            Lex.putIntent(intent, (err, data) => callback(null));
-        })
-        )
+            // delete any previously existing Holibot intents (we're going to recreate them)
+            .concat(
+            intents.map((intent) => (callback) => {
+                try {
+                    Lex.deleteIntent({ name: intent.name }, (err, data) => callback(null));
+                }
+                catch (error) {
+                    console.log('Could not delete ' + intent.name + ': ' + error);
+                }
+            })
+            )
 
-        .concat([
+            // create each intent
+            .concat(
+            intents.map((intent) => (callback) => {
+                console.log("Creating intent: " + intent.name);
+                Lex.putIntent(intent, (err, data) => callback(null));
+            })
+            )
 
-            // add permissions for Lex to call Lambda
-            addPermission,
+            .concat([
 
-            // delete the previously existing bot alias
-            deleteBotAlias,
+                // add permissions for Lex to call Lambda
+                addPermission,
 
-            // delete the previously existing bot
-            deleteBot,
+                // delete the previously existing bot alias
+                deleteBotAlias,
 
-            // create the bot
-            createBot,
+                // delete the previously existing bot
+                deleteBot,
 
-            // wait until AWS has finished building the bot
-            waitForBotProvision,
+                // create the bot
+                createBot,
 
-            // create the alias
-            createBotAlias
-        ]), (err, results) => {
-            if (err) {
+                // wait until AWS has finished building the bot
+                waitForBotProvision,
 
-                console.log("There was an error executing one of the setup steps: " + err);
-                process.exit();
-            } else {
-                console.log("All done. Now you need to do the following:");
-                console.log("1. Follow the AWS instructions to create an app in Slack: http://docs.aws.amazon.com/lex/latest/dg/slack-bot-assoc-create-app.html");
-                console.log("2. Then follow these steps to add a channel association to Slack in AWS: http://docs.aws.amazon.com/lex/latest/dg/slack-bot-assoc-create-assoc.html");
-                console.log("   As well as the Slack permissions mentioned in the AWS instructions, you need to add users:read and users:read.email.");
-                console.log("3. Visit https://api.slack.com/apps/");
-                console.log("4. Select the app you created and click Manage Distribution > Add to Slack");
-                console.log("5. Capture the OAuth access token from Slack and find your Lambda function in the AWS console");
-                console.log("6. Add an environment variable to the Lambda function called SlackToken containing the OAuth access token value from Slack.");
-                console.log("You should be good to start chatting!");
-            }
-        });
+                // create the alias
+                createBotAlias
+            ]), (err, results) => {
+                if (err) {
+
+                    console.log("There was an error executing one of the setup steps: " + err);
+                    process.exit();
+                } else {
+                    console.log("All done. Now you need to do the following:");
+                    console.log("1. Follow the AWS instructions to create an app in Slack: http://docs.aws.amazon.com/lex/latest/dg/slack-bot-assoc-create-app.html");
+                    console.log("2. Then follow these steps to add a channel association to Slack in AWS: http://docs.aws.amazon.com/lex/latest/dg/slack-bot-assoc-create-assoc.html");
+                    console.log("   As well as the Slack permissions mentioned in the AWS instructions, you need to add users:read and users:read.email.");
+                    console.log("3. Visit https://api.slack.com/apps/");
+                    console.log("4. Select the app you created and click Manage Distribution > Add to Slack");
+                    console.log("5. Capture the OAuth access token from Slack and find your Lambda function in the AWS console");
+                    console.log("6. Add an environment variable to the Lambda function called SlackToken containing the OAuth access token value from Slack.");
+                    console.log("You should be good to start chatting!");
+                }
+            });
 }
 
 /*
 Execute an sls deploy for the project to package up and deploy the Lambda function.
 */
 function deployServerlessProject(callback) {
-    console.log("Deploying Serverless project to AWS. This will create a stack in us-east-1 named holibot. Note your provided token are saved to ./token.yml. You may wish to delete them."); run("sls.cmd", ["deploy", "--stage", "prod", "--region", "us-east-1"]);
+    console.log("Deploying Serverless project to AWS. This will create a stack in us-east-1 named holibot. Note your provided token are saved to ./token.yml. You may wish to delete them.");
+
+    if (!program.refreshIntentsOnly) {
+        run("sls.cmd", ["deploy", "--stage", "prod", "--region", "us-east-1"]);
+    }
     callback(null);
 }
 
@@ -161,35 +168,45 @@ function getLambdaFunctionArn(callback) {
 Create the bot
 */
 function createBot(callback) {
-    console.log("Creating bot");
-    // get the bot object, which starts out in the local ./lex-objects/bot.json file but needs
-    // to be added to depending on which intents we are creating
-    var bot = JSON.parse(fs.readFileSync('./lex-objects/bot.json', 'utf-8'));
-    intents.forEach((intent) => {
-        bot.intents.push({
-            "intentVersion": "$LATEST",
-            "intentName": intent.name
+    if (!program.refreshIntentsOnly) {
+        console.log("Creating bot");
+        // get the bot object, which starts out in the local ./lex-objects/bot.json file but needs
+        // to be added to depending on which intents we are creating
+        var bot = JSON.parse(fs.readFileSync('./lex-objects/bot.json', 'utf-8'));
+        intents.forEach((intent) => {
+            bot.intents.push({
+                "intentVersion": "$LATEST",
+                "intentName": intent.name
+            });
         });
-    });
-    Lex.putBot(bot, (err, data) => {
+        Lex.putBot(bot, (err, data) => {
+            callback(null);
+        });
+    } else {
+        console.log("Skipping bot creation as --refresh-intents-only was specified");
         callback(null);
-    });
+    }
 }
 
 /*
 Delete an existing 'prod' alias for Holibot if it exists
 */
 function deleteBotAlias(callback) {
-    console.log("Deleting previous Holibot prod alias if necessary");
+    if (!program.refreshIntentsOnly) {
+        console.log("Deleting previous Holibot prod alias if necessary");
 
-    try {
-        Lex.deleteBotAlias({ name: 'prod', botName: 'Holibot' }, () => {
-            console.log("Bot alias deleted");
+        try {
+            Lex.deleteBotAlias({ name: 'prod', botName: 'Holibot' }, () => {
+                console.log("Bot alias deleted");
+                callback(null);
+            });
+        }
+        catch (error) {
+            console.log("Error deleting previous alias: " + err);
             callback(null);
-        });
-    }
-    catch (error) {
-        console.log("Error deleting previous alias: " + err);
+        }
+    } else {
+        console.log("Skipping deletion of previous bot alias as --refresh-intents-only.");
         callback(null);
     }
 }
@@ -198,36 +215,46 @@ function deleteBotAlias(callback) {
 Delete an existing Holibot if it exists
 */
 function deleteBot(callback) {
-    console.log("Deleting previous Holibot if necessary");
+    if (!program.refreshIntentsOnly) {
+        console.log("Deleting previous Holibot if necessary");
 
-    // delete existing bot and intents if they exist
-    Lex.getBot({ name: "Holibot", versionOrAlias: "$LATEST" },
-        (err, existingBot) => {
-            if (existingBot) {
-                console.log("Holibot already exists - deleting");
-                Lex.deleteBot({ name: existingBot.name },
-                    (err, data) => {
-                        console.log("existing bot deleted");
-                        callback(null);
-                    });
-            } else {
-                callback(null);
-            }
-        });
+        // delete existing bot and intents if they exist
+        Lex.getBot({ name: "Holibot", versionOrAlias: "$LATEST" },
+            (err, existingBot) => {
+                if (existingBot) {
+                    console.log("Holibot already exists - deleting");
+                    Lex.deleteBot({ name: existingBot.name },
+                        (err, data) => {
+                            console.log("existing bot deleted");
+                            callback(null);
+                        });
+                } else {
+                    callback(null);
+                }
+            });
+    } else {
+        console.log("Skipping deletion of previous bot as --refresh-intents-only was specified.");
+        callback(null);
+    }
 }
 
 /*
 Grant permissions for Lex to invoke the Lambda handler
 */
 function addPermission(callback) {
-    Lambda.addPermission({
-        FunctionName: holibotFunctionHandler,
-        Action: "lambda:InvokeFunction",
-        Principal: "lex.amazonaws.com",
-        StatementId: holibotFunctionHandler + "-" + makeid(5)
-    }, (err, data) => {
-        callback(null);
-    });
+    if (!program.refreshIntentsOnly) {
+        console.log("Adding permission for " + holibotFunctionHandler + " to be invokeable by Lex.");
+        Lambda.addPermission({
+            FunctionName: holibotFunctionHandler,
+            Action: "lambda:InvokeFunction",
+            Principal: "lex.amazonaws.com",
+            StatementId: holibotFunctionHandler + "-" + makeid(5)
+        }, (err, data) => {
+            callback(null);
+        });
+    } else {
+        console.log("Skipping adding permission for the Lambda handler as --refresh-intents-only was specified.");
+    }
 }
 
 /*
@@ -247,34 +274,44 @@ function makeid(n) {
 Keep checking for the existence of our bot and then call the supplied callback
 */
 function waitForBotProvision(callback) {
-    setTimeout(() => {
-        var botCreated = Lex.getBot({
-            name: 'Holibot',
-            versionOrAlias: '$LATEST'
-        }, (err, bot) => {
-            console.log("Waiting for Holibot to be provisioned before creating the alias");
-            
-            if (err) {
-                waitForBotProvision(callback);
-            } else {
-                // bot created, continue
-                callback(null);
-            }
-        });
-    }, 2000);
-}
+    if (!program.refreshIntentsOnly) {
+        setTimeout(() => {
+            var botCreated = Lex.getBot({
+                name: 'Holibot',
+                versionOrAlias: '$LATEST'
+            }, (err, bot) => {
+                console.log("Waiting for Holibot to be provisioned before creating the alias");
 
+                if (err) {
+                    waitForBotProvision(callback);
+                } else {
+                    // bot created, continue
+                    callback(null);
+                }
+            });
+        }, 2000);
+    } else {
+        console.log("--refresh-intents-only was specified so need to wait for a bot provisioning operation to happen.");
+        callback(null);
+    }
+}
 
 /*
 Create the 'prod' alias for the bot
 */
 function createBotAlias(callback) {
-    Lex.putBotAlias({
-        name: 'prod',
-        botName: 'Holibot',
-        botVersion: '$LATEST'
-    }, (err, data) => {
-        console.log("Bot alias 'prod' created.");
+    if (!program.refreshIntentsOnly) {
+        console.log("Adding 'prod' alias for Holibot.");
+        Lex.putBotAlias({
+            name: 'prod',
+            botName: 'Holibot',
+            botVersion: '$LATEST'
+        }, (err, data) => {
+            console.log("Bot alias 'prod' created.");
+            callback(null);
+        });
+    } else {
+        console.log("--refresh-intents-only was specified, so we're not creating a bot alias.");
         callback(null);
-    });
+    }
 }
